- Switch util.AddressToNodeID to return a non-negative int64 derived
from the low 64 bits of keccak256(address) (masked to 63 bits).
- Add fail-fast validation to prevent silent overwrites from:
  - duplicate operator addresses
  - derived nodeID collisions
-Reuse the validator in both fetchCurrentOperators and createSession.
-Add unit tests for validateOperatorSetNoNodeIDCollisions (incl. forced
collision via a small test seam).

Why
- Prevent subtle bugs from uint64 -> int overflow/negative IDs and guard
against the (extremely unlikely) but catastrophic case of nodeID
collisions causing split-brain or misrouting in protocol session state.

+++ pkg/bls/polynomial.go
@@ -21,7 +21,7 @@ func EvaluatePolynomial(poly polynomial.Polynomial, x int64) *fr.Element {
 // ComputeLagrangeCoefficient computes the Lagrange coefficient for participant i at x=0
 // audit: should participants register a random element than using their index ?
 // TODO(anup): there is an optimized cached version that can be done here.
-func ComputeLagrangeCoefficient(i int, participants []int) *fr.Element {
+func ComputeLagrangeCoefficient(i int64, participants []int64) *fr.Element {
 	numerator := new(fr.Element).SetOne()
 	denominator := new(fr.Element).SetOne()
 
@@ -49,8 +49,8 @@ func ComputeLagrangeCoefficient(i int, participants []int) *fr.Element {
 }
 
 // RecoverSecret recovers the secret from shares using Lagrange interpolation
-func RecoverSecret(shares map[int]*fr.Element) (*fr.Element, error) {
-	participants := make([]int, 0, len(shares))
+func RecoverSecret(shares map[int64]*fr.Element) (*fr.Element, error) {
+	participants := make([]int64, 0, len(shares))
 	for id := range shares {
 		participants = append(participants, id)
 	}
