+++ pkg/persistence/redis/redis_test.go
@@ -63,7 +63,7 @@ func TestRedisPersistence_SaveAndLoadKeyShare(t *testing.T) {
 			{CompressedBytes: []byte{1, 2, 3, 4}},
 		},
 		IsActive:       true,
-		ParticipantIDs: []int{1, 2, 3},
+		ParticipantIDs: []int64{1, 2, 3},
 	}
 
 	// Save
@@ -114,7 +114,7 @@ func TestRedisPersistence_DeleteKeyShare(t *testing.T) {
 		PrivateShare:   &privateShare,
 		Commitments:    []types.G2Point{},
 		IsActive:       true,
-		ParticipantIDs: []int{1},
+		ParticipantIDs: []int64{1},
 	}
 	err := rp.SaveKeyShareVersion(version)
 	require.NoError(t, err)
@@ -156,7 +156,7 @@ func TestRedisPersistence_ListKeyShareVersions(t *testing.T) {
 			PrivateShare:   &privateShare,
 			Commitments:    []types.G2Point{},
 			IsActive:       i == 4,
-			ParticipantIDs: []int{i},
+			ParticipantIDs: []int64{int64(i)},
 		}
 		err := rp.SaveKeyShareVersion(version)
 		require.NoError(t, err)
@@ -256,14 +256,14 @@ func TestRedisPersistence_ProtocolSessions(t *testing.T) {
 		Phase:             2,
 		StartTime:         1234567800,
 		OperatorAddresses: []string{"0x1234", "0x5678"},
-		Shares: map[int]string{
+		Shares: map[int64]string{
 			1: "share1",
 			2: "share2",
 		},
-		Commitments: map[int][]types.G2Point{
+		Commitments: map[int64][]types.G2Point{
 			1: {{CompressedBytes: []byte{1, 2, 3}}},
 		},
-		Acknowledgements: map[int]map[int]*types.Acknowledgement{
+		Acknowledgements: map[int64]map[int64]*types.Acknowledgement{
 			1: {
 				2: {PlayerID: 2, DealerID: 1, Epoch: sessionTS},
 			},
@@ -323,9 +323,9 @@ func TestRedisPersistence_DeleteProtocolSession(t *testing.T) {
 		Phase:             1,
 		StartTime:         100,
 		OperatorAddresses: []string{"0x1"},
-		Shares:            map[int]string{},
-		Commitments:       map[int][]types.G2Point{},
-		Acknowledgements:  map[int]map[int]*types.Acknowledgement{},
+		Shares:            map[int64]string{},
+		Commitments:       map[int64][]types.G2Point{},
+		Acknowledgements:  map[int64]map[int64]*types.Acknowledgement{},
 	}
 	err := rp.SaveProtocolSession(session)
 	require.NoError(t, err)
@@ -354,9 +354,9 @@ func TestRedisPersistence_ListProtocolSessions(t *testing.T) {
 			Phase:             1,
 			StartTime:         baseTS + int64(i*100),
 			OperatorAddresses: []string{},
-			Shares:            map[int]string{},
-			Commitments:       map[int][]types.G2Point{},
-			Acknowledgements:  map[int]map[int]*types.Acknowledgement{},
+			Shares:            map[int64]string{},
+			Commitments:       map[int64][]types.G2Point{},
+			Acknowledgements:  map[int64]map[int64]*types.Acknowledgement{},
 		}
 		err := rp.SaveProtocolSession(session)
 		require.NoError(t, err)
@@ -453,7 +453,7 @@ func TestRedisPersistence_ThreadSafety(t *testing.T) {
 					PrivateShare:   &privateShare,
 					Commitments:    []types.G2Point{},
 					IsActive:       false,
-					ParticipantIDs: []int{id},
+					ParticipantIDs: []int64{int64(id)},
 				}
 				err := rp.SaveKeyShareVersion(version)
 				assert.NoError(t, err)
